fix(nrL1_stats): Do not start/stop/merge timers on mixed slots
When timers are started/stopped/merged on mixed slot,
the measured times are not homogeneous as the mixed slots are much
lighter than full UL/DL slots.
Then, to make the analysis of results easier,
we do not measure the mixed slots.
We also do not start ULSCH decoding timer in cases of missing transmission.

Signed-off-by: Romain Beurdouche <romain.beurdouche@eurecom.fr>